feat: ModelStrategy complete with test cases - test still failing
