+++ src/models/__tests__/single-test.json
@@ -1,0 +1,21 @@
+[
+  [
+   {
+      "ageInMonths": 4.3,
+      "muac": 130,
+      "spo2": 96,
+      "weight": 7,
+      "sex": "Male",
+      "illnessDuration": "48h-7d",
+      "bulgingFontanelle": false,
+      "isNeonate": false,
+      "neonatalJaundice": false,
+      "feedingWell": true
+    },
+    {
+      "modelName": "Model 0-6C",
+      "riskCategory": "Moderate",
+      "riskScore": 3.26
+    }
+  ]
+]
